Updated version as a pip package
