--- exemple.html
@@ -1,68 +1,0 @@
-<!-- CECI EST LE TREE HTML D'UNE RECHERCHE TYPIQUE GOOGLE SCHOLAR 
-AVEC L'ENDROIT OU TROUVER LES INFORMATIONS -->
-
-<div id="gs_res_ccl"> <!-- Balise générale d'informations -->
-
-	<div id="gs_res_ccl_mid"> <!-- Les articles -->
-
-		<div class="gs_r gs_or gs_scl">  <!-- Un article -->
-			
-
-
-			<div class="gs_ggs gs_fl">  <!-- Lien pour télécharger full text -->
-				<div class="gs_ggsd"> <div class="gs_or_ggsm">
-
-					<a href="lien pour télécharger">
-						<span class="gs_ctg2"> PDF ou HTML.. </span>
-						Nom du site 
-					</a>
-
-			</div></div></div>
-
-
-
-			<div class="gs_ri">  <!-- Description article -->
-				<h3 class="gs_rt">
-					<a href="lien vers article"> Titre article </a>
-				</h3>
-
-				<div class="gs_a">
-					Soit des a ou text avec :
-					auteur, auteur, auteur.. - journal, année - site/librairie
-				</div>
-
-				<div class="gs_rs">
-					Texte résumant le début de l'article sur la page google
-				</div>
-
-				<div class="gs_fl">
-					<a href=".."> Ajouter aux favoris </a>
-					<a href=".."> Citer </a>
-					<a href=".."> Cité n fois </a>
-					...
-				</div>
-			</div>
-	
-	</div></div>
-
-	<div id="gs_res_ccl_bot"> <!-- Pagination en bas de la page -->
-		<div id="gs_n">
-			<center>
-				<table ..>
-					<td></td>
-					<td></td>
-					...
-					<td align="left" nowrap>
-						
-						<!-- Si il reste une page : -->
-						<a href="lien page suivante"/>
-						
-						<!-- Si c'est la fin -->
-						<span class="gs_ico gs_ico_nav_last"/>
-
-					</td>
-				</table>
-			</center>
-	</div></div>
-	
-</div>
