+++ setup.cfg
@@ -1,0 +1,43 @@
+[metadata]
+name = rst
+version = 0.0.1
+description = A tool to scrape research in academic databases like Google Scholar
+long_description = file: README.md, LICENSE
+keywords = scraper, google scholar, CNKI, research
+license = GNU General Public License v3.0
+classifiers =
+    Development Status :: 2 - Pre-Alpha
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
+    Natural Language :: English
+    Operating System :: OS Independent
+    Topic :: Scientific/Engineering :: Information Analysis
+    Programming Language :: Python :: 3
+
+[options]
+zip_safe = False
+include_package_data = True
+packages = find:
+install_requires =
+    requests
+    beautifulsoup4
+    selenium
+    fake-useragent
+    pathlib
+    importlib-metadata; python_version<"3.8"
+
+[options.package_data]
+* = *.txt, *.md
+
+[options.entry_points]
+console_scripts =
+    executable-name = rst.module:function
+
+[options.extras_require]
+
+[options.packages.find]
+exclude =
+    examples*
+    tools*
+    docs*
+    rst.tests*
