+++ setup.py
@@ -1,0 +1,41 @@
+from setuptools import setup, find_packages
+
+if __name__ == '__main__':
+    setup()
+
+#import os
+
+
+#def read(filename):
+#    filepath = os.path.join(os.path.dirname(__file__), filename)
+#    file = open(filepath, 'r')
+#    return file.read()
+
+
+#setup(
+#    name='Research Scraping Tool',
+#    version='0.0.1',
+#    author='Ben Mueller',
+#    author_email='bmueller@madwonk.com',
+#    description='A tool to scrape research in academic databases like Google Scholar',
+#    long_description=read('README.md'),
+#    long_description_content_type="text/markdown",
+#    license='GNU General Public License v3.0',
+#    keywords=[],
+#    url='https://github.com/Madwonk/Research-Scraping-Tool',
+#    packages=find_packages(),
+#    scripts=[],
+#    install_requires=['requests', 'beautifulsoup4', 'selenium', 'fake_useragent', 'pathlib'],
+#    include_package_data=True,
+#    zip_safe=False,
+#    classifiers=[
+#        'Development Status :: 2 - Pre-Alpha',
+#        'Environment :: Console',
+#        'Intended Audience :: Science/Research',
+#        'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
+#        'Natural Language :: English',
+#        'Operating System :: OS Independent',
+#        'Programming Language :: Python :: 3',
+#        'Topic :: Scientific/Engineering :: Information Analysis'
+#    ],
+#)
